requirements.txt
albumentations==1.3.0
babel==2.14.0
beautifulsoup4==4.12.3
black==23.1.0
build==1.2.1
cachetools==5.3.3
colorama==0.4.6
colour==0.1.5
einops==0.6.1
evalutils==0.4.2
flake8==6.0.0
flake8-html==0.4.3
flatbuffers==24.3.7
fonttools==4.50.0
future==0.18.2
genbadge==1.1.0
geojson==3.0.0
h5py==3.9.0
huggingface-hub==0.22.2
ipykernel==6.29.3
ipython==8.22.2
ipython-genutils==0.2.0
jedi==0.19.1
jupyterlab==4.0.2
kaleido==0.2.1
keras==2.12.0
mako==1.2.4
matplotlib==3.7.1
natsort==8.4.0
numba==0.59.0
nvitop==1.3.2
opencv-python-headless==4.5.4.58
openslide-python==1.3.1
opt-einsum==3.3.0
pandarallel==1.6.5
pandas==1.4.3
pathopatch==1.0.2
pip-tools==7.4.1
pre-commit==3.0.3
pyaml==24.7.0
pycm==4.0
pydantic==1.10.4
pydantic-compat==0.1.2
pydantic-core==2.20.1
pydicom==2.4.4
python-snappy==0.7.3
pyvips==2.2.3
pyjwt==2.6.0
rasterio==1.3.5.post1
ray==2.9.3
ruff==0.5.6
schema==0.7.5
scikit-base==0.7.8
scikit-image==0.19.3
scikit-learn==1.3.0
scikit-survival==0.22.2
scipy==1.8.1
seaborn==0.12.2
setuptools==65.6.3
shapely==1.8.5.post1
simpleitk==2.2.1
snappy==3.1.1
tabulate==0.9.0
tensorboard==2.12.3
tensorflow==2.12.0
termcolor==2.4.0
timm==1.0.8
torchinfo==1.8.0
torchmetrics==0.11.4
torchstain==1.3.0
tqdm==4.65.0
ujson==5.8.0
wandb==0.15.4
wsidicom==0.20.4
wsidicomizer==0.14.1
xgboost==2.1.1
